Fixed issue with fieldgroup expanded and collapsed not working.  It turned out we had a conflict in how the display property was being used on field groups.  I had mistakenly started using it for inline/block in forms.
This may be a little confusing.  I have changed it so the following now works:
collapsed=1  - Field group is collapsed by default on the edit form
display=collapsed  - Field group is collapsed by default in the view tab
display_style=inline - Field group fields are displayed inline in the form.
display_style=block  - Field group fields are displayed in block form in the form.

This change will break the previous change that allowed you to use display=inline or display=block to indicate field flow in the form.