Update RecordBrowser.js
Hello, it's my first correction on github. Sorry for my poor english, I'm french.
Inside code of "$.fn.RecordBrowserWidget", functions updatePermissions() and updateEditable() are declared only if the object isn't frozen but later in the code, frozen or not, if the object has a value these two functions are called causing an exception.
So i move the declaration of those two functions before.

This bug can be reproduced in a edit form with two lookup fields witch have filled values but in readonly mode (using getPermissions for example) : the second field doesn't convert the value of the lookup and let the id of the field

Thanks for your job. I love Xataface !