DE Translation for g2 module
The translation keys of the g2 module are missing in the standard de.ini and in the spreadsheet on https://docs.google.com/spreadsheets/d/1qkurdRSHAQPKvkcVcdDudEOAQnVxKpz4L8JD1P2xpe4

As stated in another pull request  (https://github.com/shannah/xataface/pull/50)  there seems to be another spreadsheet. But I do not have access to any of them.

So here is another proposal for the de.ini of the g2 module.