Added support for making fields dependent on other fields, and automatically load values via ajax without the need for javascript. Usage:
The above example would make the field dependent on the ProgramID field.  Whenever ProgramID changed, the field would fetch a new value from the provided URL except replacing {ProgramID} with the value of the ProgramID field on the form.
The query after the hash describes how to get the value from the web service using [jsonPath](http://jsonpath.com/) syntax.