Added support for explicit case-sensitive and case-insensitive search on a column either by specifying the collate property of the fields.ini or using the new URL query convention.
Using fields.ini you use the full collation you want searches to be done on as follows:
collate=utf8_unicode_ci

If performing a search in the web interface, you can now add 'commands' to the search with the following notation:

~#query#commands#

e.g. prefix with '~#', and suffix with '#', and split the query from the commands with a #.  Right now there are only 2 commands: ci and cs.  ci is case insensitive, cs is case-sensitive.

Examples:  Searching for the phrase 'about' case sensitive:

~#about#cs#

Searching for the phrase 'about' in a case-insensitive way:

~#about#ci#